+++ iacservice-20210806/alibabacloud_iacservice20210806/models/_create_parameter_set_request.py
@@ -72,11 +72,13 @@ class CreateParameterSetRequestParameters(DaraModel):
     def __init__(
         self,
         name: str = None,
+        secret: bool = None,
         status: str = None,
         type: str = None,
         value: str = None,
     ):
         self.name = name
+        self.secret = secret
         self.status = status
         self.type = type
         self.value = value
@@ -91,6 +93,9 @@ class CreateParameterSetRequestParameters(DaraModel):
             result = _map
         if self.name is not None:
             result['name'] = self.name
+
+        if self.secret is not None:
+            result['secret'] = self.secret
 
         if self.status is not None:
             result['status'] = self.status
@@ -107,6 +112,9 @@ class CreateParameterSetRequestParameters(DaraModel):
         m = m or dict()
         if m.get('name') is not None:
             self.name = m.get('name')
+
+        if m.get('secret') is not None:
+            self.secret = m.get('secret')
 
         if m.get('status') is not None:
             self.status = m.get('status')