+++ iacservice-20210806/alibabacloud_iacservice20210806/models/_list_parameter_sets_response_body.py
@@ -213,11 +213,13 @@ class ListParameterSetsResponseBodyParameterSetsParameters(DaraModel):
     def __init__(
         self,
         name: str = None,
+        secret: bool = None,
         status: str = None,
         type: str = None,
         value: Any = None,
     ):
         self.name = name
+        self.secret = secret
         self.status = status
         self.type = type
         self.value = value
@@ -232,6 +234,9 @@ class ListParameterSetsResponseBodyParameterSetsParameters(DaraModel):
             result = _map
         if self.name is not None:
             result['name'] = self.name
+
+        if self.secret is not None:
+            result['secret'] = self.secret
 
         if self.status is not None:
             result['status'] = self.status
@@ -248,6 +253,9 @@ class ListParameterSetsResponseBodyParameterSetsParameters(DaraModel):
         m = m or dict()
         if m.get('name') is not None:
             self.name = m.get('name')
+
+        if m.get('secret') is not None:
+            self.secret = m.get('secret')
 
         if m.get('status') is not None:
             self.status = m.get('status')