+++ iacservice-20210806/alibabacloud_iacservice20210806/models/_update_parameter_set_attribute_request.py
@@ -63,11 +63,13 @@ class UpdateParameterSetAttributeRequestParameters(DaraModel):
     def __init__(
         self,
         name: str = None,
+        secret: bool = None,
         status: str = None,
         type: str = None,
         value: str = None,
     ):
         self.name = name
+        self.secret = secret
         self.status = status
         self.type = type
         self.value = value
@@ -82,6 +84,9 @@ class UpdateParameterSetAttributeRequestParameters(DaraModel):
             result = _map
         if self.name is not None:
             result['name'] = self.name
+
+        if self.secret is not None:
+            result['secret'] = self.secret
 
         if self.status is not None:
             result['status'] = self.status
@@ -98,6 +103,9 @@ class UpdateParameterSetAttributeRequestParameters(DaraModel):
         m = m or dict()
         if m.get('name') is not None:
             self.name = m.get('name')
+
+        if m.get('secret') is not None:
+            self.secret = m.get('secret')
 
         if m.get('status') is not None:
             self.status = m.get('status')